+++ CHANGELOG.md
@@ -2,11 +2,11 @@
 
 * **Send logging messages to Telegram (experimental)**
 * **Support tower full skip for the 130th level**
+* **Switch response classes to `pydantic`**
 * Improve packaging, remove `--log-file` option and improve Docker integration
 * Use `ujson` instead of `json`
 * Update resources
 * Use `loguru` for logging and improve testing
-* Move `class Letter` to `dataclasses_`
 
 # `2.3`